[Feat] 경험치 시스템 구현 / BookInfo 응답 수정 / 클라이밍 감상평 조회 API 수정

+++ core/src/main/java/org/bookwoori/core/domain/member/dto/response/MemberResponseDto.java
@@ -11,7 +11,7 @@ public record MemberResponseDto(
     String backgroundImg,
     int level,
     String mountain,
-    int height,
+    double height,
     int totalPage) {
 
     public static MemberResponseDto from(Member member, boolean isMine) {
